+++ sdk/terraform/arm-terraform/src/api/terraform/index.ts
@@ -40,7 +40,7 @@ export function _terraformExportTerraformSend(
 export async function _terraformExportTerraformDeserialize(
   result: PathUncheckedResponse,
 ): Promise<TerraformOperationStatus> {
-  const expectedStatuses = ["202", "200"];
+  const expectedStatuses = ["202", "202", "200"];
   if (!expectedStatuses.includes(result.status)) {
     throw createRestError(result);
   }
@@ -55,10 +55,16 @@ export function terraformExportTerraform(
   body: BaseExportModelUnion,
   options: TerraformExportTerraformOptionalParams = { requestOptions: {} },
 ): PollerLike<OperationState<TerraformOperationStatus>, TerraformOperationStatus> {
-  return getLongRunningPoller(context, _terraformExportTerraformDeserialize, ["202", "200"], {
-    updateIntervalInMs: options?.updateIntervalInMs,
-    abortSignal: options?.abortSignal,
-    getInitialResponse: () => _terraformExportTerraformSend(context, subscriptionId, body, options),
-    resourceLocationConfig: "azure-async-operation",
-  }) as PollerLike<OperationState<TerraformOperationStatus>, TerraformOperationStatus>;
+  return getLongRunningPoller(
+    context,
+    _terraformExportTerraformDeserialize,
+    ["202", "202", "200"],
+    {
+      updateIntervalInMs: options?.updateIntervalInMs,
+      abortSignal: options?.abortSignal,
+      getInitialResponse: () =>
+        _terraformExportTerraformSend(context, subscriptionId, body, options),
+      resourceLocationConfig: "azure-async-operation",
+    },
+  ) as PollerLike<OperationState<TerraformOperationStatus>, TerraformOperationStatus>;
 }