+++ sdk/terraform/arm-terraform/src/index.ts
@@ -17,6 +17,8 @@ export {
   KnownTargetProvider,
   TargetProvider,
   ExportQuery,
+  KnownAuthorizationScopeFilter,
+  AuthorizationScopeFilter,
   ExportResource,
   ExportResourceGroup,
   ErrorDetail,