+++ sdk/terraform/arm-terraform/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Release History
     
+## 1.0.0-beta.2 (2025-01-15)
+Compared with version 1.0.0-beta.1
+    
+### Features Added
+
+  - Added Type Alias AuthorizationScopeFilter
+  - Interface ExportQuery has a new optional parameter authorizationScopeFilter
+  - Interface ExportQuery has a new optional parameter table
+  - Interface ExportResult has a new optional parameter import
+  - Added Enum KnownAuthorizationScopeFilter
+  - Enum KnownOrigin has a new value "user,system"
+  - Enum KnownOrigin has a new value system
+  - Enum KnownOrigin has a new value user
+
+### Breaking Changes
+
+  - Enum KnownOrigin no longer has value System
+  - Enum KnownOrigin no longer has value User
+  - Enum KnownOrigin no longer has value UserSystem
+    
+    
 ## 1.0.0-beta.1 (2024-11-18)
 
 ### Features Added